# Allow IncrementalIndex to store Long/Float dimensions