--- a/processing/src/main/java/io/druid/query/filter/ValueMatcherFactory.java
+++ b/processing/src/main/java/io/druid/query/filter/ValueMatcherFactory.java
@@ -26,7 +26,7 @@ import com.metamx.collections.spatial.search.Bound;
  */
 public interface ValueMatcherFactory
 {
-  public ValueMatcher makeValueMatcher(String dimension, String value);
-  public ValueMatcher makeValueMatcher(String dimension, Predicate<String> value);
+  public ValueMatcher makeValueMatcher(String dimension, Comparable value);
+  public ValueMatcher makeValueMatcher(String dimension, Predicate value);
   public ValueMatcher makeValueMatcher(String dimension, Bound bound);
 }