--- a/processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
+++ b/processing/src/main/java/io/druid/segment/incremental/IncrementalIndex.java
@@ -24,6 +24,7 @@ import com.google.common.base.Function;
 import com.google.common.base.Strings;
 import com.google.common.base.Supplier;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -31,6 +32,7 @@ import com.google.common.primitives.Ints;
 import com.google.common.primitives.Longs;
 import com.metamx.common.IAE;
 import com.metamx.common.ISE;
+import com.metamx.common.parsers.ParseException;
 import io.druid.data.input.InputRow;
 import io.druid.data.input.MapBasedRow;
 import io.druid.data.input.Row;
@@ -78,6 +80,76 @@ import java.util.concurrent.atomic.AtomicInteger;
 public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>, Closeable
 {
   private volatile DateTime maxIngestedEventTime;
+
+  private static final Map<String, ValueType> TYPE_MAP = ImmutableMap.<String, ValueType>builder()
+      .put("Long[]", ValueType.LONG)
+      .put("Double[]", ValueType.FLOAT)
+      .put("String[]", ValueType.STRING)
+      .put("Long", ValueType.LONG)
+      .put("Double", ValueType.FLOAT)
+      .put("String", ValueType.STRING)
+      .build();
+
+  private static final Function<Object, String> STRING_TRANSFORMER = new Function<Object, String>()
+  {
+    @Override
+    public String apply(final Object o)
+    {
+      return String.valueOf(o);
+    }
+  };
+
+  private static final Function<Object, Long> LONG_TRANSFORMER = new Function<Object, Long>()
+  {
+    @Override
+    public Long apply(final Object o)
+    {
+      if (o == null) {
+        return null;
+      }
+      if (o instanceof String) {
+        try {
+          return Long.valueOf((String) o);
+        }
+        catch (NumberFormatException nfe) {
+          throw new ParseException(nfe, "Unable to parse value[%s] as long in column: ", o);
+        }
+      }
+      if (o instanceof Number) {
+        return ((Number) o).longValue();
+      }
+      return null;
+    }
+  };
+
+  private static final Function<Object, Float> FLOAT_TRANSFORMER = new Function<Object, Float>()
+  {
+    @Override
+    public Float apply(final Object o)
+    {
+      if (o == null) {
+        return null;
+      }
+      if (o instanceof String) {
+        try {
+          return Float.valueOf((String) o);
+        }
+        catch (NumberFormatException nfe) {
+          throw new ParseException(nfe, "Unable to parse value[%s] as float in column: ", o);
+        }
+      }
+      if (o instanceof Number) {
+        return ((Number) o).floatValue();
+      }
+      return null;
+    }
+  };
+
+  private static final Map<ValueType, Function> VALUE_TRANSFORMS = ImmutableMap.<ValueType, Function>builder()
+      .put(ValueType.LONG, LONG_TRANSFORMER)
+      .put(ValueType.FLOAT, FLOAT_TRANSFORMER)
+      .put(ValueType.STRING, STRING_TRANSFORMER)
+      .build();
 
   public static ColumnSelectorFactory makeColumnSelectorFactory(
       final AggregatorFactory agg,
@@ -271,6 +343,7 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
   private final Metadata metadata;
 
   private final Map<String, MetricDesc> metricDescs;
+
   private final Map<String, DimensionDesc> dimensionDescs;
   private final Map<String, ColumnCapabilitiesImpl> columnCapabilities;
   private final List<DimDim> dimValues;
@@ -347,19 +420,32 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     }
   }
 
-  private DimDim newDimDim(String dimension)
-  {
-    return new NullValueConverterDimDim(makeDimDim(dimension, dimensionDescs));
+  private DimDim newDimDim(String dimension, ValueType type) {
+    DimDim newDimDim;
+    switch (type) {
+      case LONG:
+        newDimDim = makeDimDim(dimension, getDimensionDescs());
+        break;
+      case FLOAT:
+        newDimDim = makeDimDim(dimension, getDimensionDescs());
+        break;
+      case STRING:
+        newDimDim = new NullValueConverterDimDim(makeDimDim(dimension, getDimensionDescs()));
+        break;
+      default:
+        throw new IAE("Invalid column type: " + type);
+    }
+    return newDimDim;
   }
+
+  // use newDimDim() to create a DimDim, makeDimDim() provides the subclass-specific implementation
+  protected abstract DimDim makeDimDim(String dimension, Object lock);
 
   public abstract ConcurrentNavigableMap<TimeAndDims, Integer> getFacts();
 
   public abstract boolean canAppendRow();
 
   public abstract String getOutOfRowsReason();
-
-  // use newDimDim
-  protected abstract DimDim makeDimDim(String dimension, Object lock);
 
   protected abstract AggregatorType[] initAggs(
       AggregatorFactory[] metrics,
@@ -407,6 +493,48 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     return row;
   }
 
+  private ValueType getTypeFromDimVal(Object dimVal)
+  {
+    Object singleVal;
+    if (dimVal instanceof List) {
+      List dimValList = (List) dimVal;
+      singleVal = dimValList.size() == 0 ? null : dimValList.get(0);
+    } else {
+      singleVal = dimVal;
+    }
+
+    if (singleVal == null) {
+      return null;
+    }
+
+    return TYPE_MAP.get(singleVal.getClass().getSimpleName());
+  }
+
+  private List<Comparable> getRowDimensionAsComparables(InputRow row, String dimension, ValueType type)
+  {
+    final Object dimVal = row.getRaw(dimension);
+    final Function transformer = VALUE_TRANSFORMS.get(type);
+    final List<Comparable> dimensionValues;
+    try {
+      if (dimVal == null) {
+        dimensionValues = Collections.emptyList();
+      } else if (dimVal instanceof List) {
+        dimensionValues = Lists.transform((List) dimVal, transformer);
+      } else {
+        dimensionValues = Collections.singletonList((Comparable) transformer.apply(dimVal));
+      }
+    }
+    catch (ParseException pe) {
+      throw new ParseException(pe.getMessage() + dimension);
+    }
+    return dimensionValues;
+  }
+
+  public Map<String, DimensionDesc> getDimensionDescs()
+  {
+    return dimensionDescs;
+  }
+
   /**
    * Adds a new row.  The row might correspond with another row that already exists, in which case this will
    * update that row instead of inserting a new one.
@@ -446,13 +574,17 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     final List<String> rowDimensions = row.getDimensions();
 
     int[][] dims;
+    ValueType[] types;
     List<int[]> overflow = null;
+    List<ValueType> overflowTypes = null;
     synchronized (dimensionDescs) {
       dims = new int[dimensionDescs.size()][];
+      types = new ValueType[dimensionDescs.size()];
       for (String dimension : rowDimensions) {
-        List<String> dimensionValues = row.getDimension(dimension);
+        List<Comparable> dimensionValues;
 
         ColumnCapabilitiesImpl capabilities;
+        ValueType valType = null;
         DimensionDesc desc = dimensionDescs.get(dimension);
         if (desc != null) {
           capabilities = desc.getCapabilities();
@@ -460,10 +592,17 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
           capabilities = columnCapabilities.get(dimension);
           if (capabilities == null) {
             capabilities = new ColumnCapabilitiesImpl();
-            capabilities.setType(ValueType.STRING);
+            // TODO: For schemaless type discovery, assume everything is a String for now, can change later.
+            //valType = getTypeFromDimVal(row.getRaw(dimension));
+            if (valType == null) {
+              valType = ValueType.STRING;
+            }
+            capabilities.setType(valType);
             columnCapabilities.put(dimension, capabilities);
           }
         }
+        valType = capabilities.getType();
+        dimensionValues = getRowDimensionAsComparables(row, dimension, valType);
 
         // Set column capabilities as data is coming in
         if (!capabilities.hasMultipleValues() && dimensionValues.size() > 1) {
@@ -475,8 +614,10 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
 
           if (overflow == null) {
             overflow = Lists.newArrayList();
+            overflowTypes = Lists.newArrayList();
           }
           overflow.add(getDimVals(desc.getValues(), dimensionValues));
+          overflowTypes.add(valType);
         } else if (desc.getIndex() > dims.length || dims[desc.getIndex()] != null) {
           /*
            * index > dims.length requires that we saw this dimension and added it to the dimensionOrder map,
@@ -490,6 +631,7 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
           throw new ISE("Dimension[%s] occurred more than once in InputRow", dimension);
         } else {
           dims[desc.getIndex()] = getDimVals(desc.getValues(), dimensionValues);
+          types[desc.getIndex()] = valType;
         }
       }
     }
@@ -497,15 +639,19 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     if (overflow != null) {
       // Merge overflow and non-overflow
       int[][] newDims = new int[dims.length + overflow.size()][];
+      ValueType[] newTypes = new ValueType[dims.length + overflow.size()];
       System.arraycopy(dims, 0, newDims, 0, dims.length);
+      System.arraycopy(types, 0, newTypes, 0, dims.length);
       for (int i = 0; i < overflow.size(); ++i) {
         newDims[dims.length + i] = overflow.get(i);
+        newTypes[dims.length + i] = overflowTypes.get(i);
       }
       dims = newDims;
+      types = newTypes;
     }
 
     long truncated = gran.truncate(row.getTimestampFromEpoch());
-    return new TimeAndDims(Math.max(truncated, minTimestamp), dims);
+    return new TimeAndDims(Math.max(truncated, minTimestamp), dims, types);
   }
 
   private synchronized void updateMaxIngestedTime(DateTime eventTime)
@@ -535,7 +681,7 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     return getFacts().lastKey().getTimestamp();
   }
 
-  private int[] getDimVals(final DimDim dimLookup, final List<String> dimValues)
+  private int[] getDimVals(final DimDim dimLookup, final List<Comparable> dimValues)
   {
     if (dimValues.size() == 0) {
       // NULL VALUE
@@ -544,10 +690,13 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     }
 
     if (dimValues.size() == 1) {
-      return new int[]{dimLookup.add(dimValues.get(0))};
+      Comparable dimVal = dimValues.get(0);
+      // For Strings, return an array of dictionary-encoded IDs
+      // For numerics, return the numeric values directly
+      return new int[]{dimLookup.add(dimVal)};
     }
 
-    String[] dimArray = dimValues.toArray(new String[dimValues.size()]);
+    Comparable[] dimArray = dimValues.toArray(new Comparable[dimValues.size()]);
     Arrays.sort(dimArray);
 
     final int[] retVal = new int[dimArray.length];
@@ -649,7 +798,7 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
   @GuardedBy("dimensionDescs")
   private DimensionDesc addNewDimension(String dim, ColumnCapabilitiesImpl capabilities)
   {
-    DimensionDesc desc = new DimensionDesc(dimensionDescs.size(), dim, newDimDim(dim), capabilities);
+    DimensionDesc desc = new DimensionDesc(dimensionDescs.size(), dim, newDimDim(dim, capabilities.getType()), capabilities);
     if (dimValues.size() != desc.getIndex()) {
       throw new ISE("dimensionDescs and dimValues for [%s] is out of sync!!", dim);
     }
@@ -724,11 +873,13 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
                 final TimeAndDims timeAndDims = input.getKey();
                 final int rowOffset = input.getValue();
 
-                int[][] theDims = timeAndDims.getDims();
+                int[][] theDims = timeAndDims.getDims(); //TODO: remove dictionary encoding for numerics later
+                ValueType[] types = timeAndDims.getTypes();
 
                 Map<String, Object> theVals = Maps.newLinkedHashMap();
                 for (int i = 0; i < theDims.length; ++i) {
                   int[] dim = theDims[i];
+                  ValueType type = types[i];
                   DimensionDesc dimensionDesc = dimensions.get(i);
                   if (dimensionDesc == null) {
                     continue;
@@ -739,13 +890,21 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
                     continue;
                   }
                   if (dim.length == 1) {
-                    theVals.put(dimensionName, Strings.nullToEmpty(dimensionDesc.getValues().getValue(dim[0])));
-                  } else {
-                    String[] dimStringValue = new String[dim.length];
-                    for (int j = 0; j < dimStringValue.length; j++) {
-                      dimStringValue[j] = Strings.nullToEmpty(dimensionDesc.getValues().getValue(dim[j]));
+                    Comparable val = dimensionDesc.getValues().getValue(dim[0]);
+                    if (type == ValueType.STRING) {
+                      val = Strings.nullToEmpty((String) val);
                     }
-                    theVals.put(dimensionName, dimStringValue);
+                    theVals.put(dimensionName, val);
+                  } else {
+                    Comparable[] dimVals = new Comparable[dim.length];
+                    for (int j = 0; j < dimVals.length; j++) {
+                      Comparable val = dimensionDesc.getValues().getValue(dim[j]);
+                      if (type == ValueType.STRING) {
+                        val = Strings.nullToEmpty((String) val);
+                      }
+                      dimVals[j] = val;
+                    }
+                    theVals.put(dimensionName, dimVals);
                   }
                 }
 
@@ -852,42 +1011,42 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     }
   }
 
-  static interface DimDim
+  static interface DimDim<T extends Comparable<? super T>>
   {
-    public int getId(String value);
+    public int getId(T value);
 
-    public String getValue(int id);
+    public T getValue(int id);
 
-    public boolean contains(String value);
+    public boolean contains(T value);
 
     public int size();
 
-    public String getMinValue();
+    public T getMinValue();
 
-    public String getMaxValue();
+    public T getMaxValue();
 
-    public int add(String value);
+    public int add(T value);
 
     public SortedDimLookup sort();
   }
 
-  static interface SortedDimLookup
+  static interface SortedDimLookup<T extends Comparable<? super T>>
   {
     public int size();
 
-    public int idToIndex(int id);
+    public int getSortedIdFromUnsortedId(int id);
 
-    public int indexToId(int index);
+    public int getUnsortedIdFromSortedId(int index);
 
-    public String getValue(int index);
+    public T getValueFromSortedId(int index);
   }
 
   /**
    * implementation which converts null strings to empty strings and vice versa.
    */
-  static class NullValueConverterDimDim implements DimDim
+  static class NullValueConverterDimDim implements DimDim<String>
   {
-    private final DimDim delegate;
+    private final DimDim<String> delegate;
 
     NullValueConverterDimDim(DimDim delegate)
     {
@@ -943,9 +1102,9 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     }
   }
 
-  private static class NullValueConverterDimLookup implements SortedDimLookup
+  private static class NullValueConverterDimLookup implements SortedDimLookup<String>
   {
-    private final SortedDimLookup delegate;
+    private final SortedDimLookup<String> delegate;
 
     public NullValueConverterDimLookup(SortedDimLookup delegate)
     {
@@ -959,21 +1118,21 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     }
 
     @Override
-    public int indexToId(int index)
+    public int getUnsortedIdFromSortedId(int index)
     {
-      return delegate.indexToId(index);
+      return delegate.getUnsortedIdFromSortedId(index);
     }
 
     @Override
-    public int idToIndex(int id)
+    public int getSortedIdFromUnsortedId(int id)
     {
-      return delegate.idToIndex(id);
+      return delegate.getSortedIdFromUnsortedId(id);
     }
 
     @Override
-    public String getValue(int index)
+    public String getValueFromSortedId(int index)
     {
-      return Strings.emptyToNull(delegate.getValue(index));
+      return Strings.emptyToNull(delegate.getValueFromSortedId(index));
     }
   }
 
@@ -981,14 +1140,17 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
   {
     private final long timestamp;
     private final int[][] dims;
+    private final ValueType[] types;
 
     TimeAndDims(
         long timestamp,
-        int[][] dims
+        int[][] dims,
+        ValueType[] types
     )
     {
       this.timestamp = timestamp;
       this.dims = dims;
+      this.types = types;
     }
 
     long getTimestamp()
@@ -999,6 +1161,11 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
     int[][] getDims()
     {
       return dims;
+    }
+
+    public ValueType[] getTypes()
+    {
+      return types;
     }
 
     @Override
@@ -1045,6 +1212,26 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
 
       int index = 0;
       while (retVal == 0 && index < numComparisons) {
+        ValueType lhsType = lhs.types[index];
+        ValueType rhsType = rhs.types[index];
+
+        if (lhsType == null) {
+          if (rhsType == null) {
+            ++index;
+            continue;
+          }
+          return -1;
+        }
+
+        if (rhsType == null) {
+          return 1;
+        }
+
+        retVal = lhsType.compareTo(rhsType);
+        if (retVal != 0) {
+          return retVal;
+        }
+
         final int[] lhsIdxs = lhs.dims[index];
         final int[] rhsIdxs = rhs.dims[index];
 
@@ -1066,8 +1253,8 @@ public abstract class IncrementalIndex<AggregatorType> implements Iterable<Row>,
         while (retVal == 0 && valsIndex < lhsIdxs.length) {
           if (lhsIdxs[valsIndex] != rhsIdxs[valsIndex]) {
             final DimDim dimLookup = dimValues.get(index);
-            final String lhsVal = dimLookup.getValue(lhsIdxs[valsIndex]);
-            final String rhsVal = dimLookup.getValue(rhsIdxs[valsIndex]);
+            final Comparable lhsVal = dimLookup.getValue(lhsIdxs[valsIndex]);
+            final Comparable rhsVal = dimLookup.getValue(rhsIdxs[valsIndex]);
             if (lhsVal != null && rhsVal != null) {
               retVal = lhsVal.compareTo(rhsVal);
             } else if (lhsVal == null ^ rhsVal == null) {