--- a/processing/src/main/java/io/druid/segment/incremental/IncrementalIndexStorageAdapter.java
+++ b/processing/src/main/java/io/druid/segment/incremental/IncrementalIndexStorageAdapter.java
@@ -234,10 +234,10 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
               {
                 cursorMap = index.getSubMap(
                     new IncrementalIndex.TimeAndDims(
-                        timeStart, new int[][]{}
+                        timeStart, new int[][]{}, null
                     ),
                     new IncrementalIndex.TimeAndDims(
-                        Math.min(actualInterval.getEndMillis(), gran.next(input)), new int[][]{}
+                        Math.min(actualInterval.getEndMillis(), gran.next(input)), new int[][]{}, null
                     )
                 );
                 if (descending) {
@@ -414,8 +414,10 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
                   @Override
                   public String lookupName(int id)
                   {
-                    final String value = dimValLookup.getValue(id);
-                    return extractionFn == null ? value : extractionFn.apply(value);
+                    // TODO: needs update to DimensionSelector interface to allow multi-types, just use Strings for now
+                    final Comparable value = dimValLookup.getValue(id);
+                    final String strValue = value == null ? null : value.toString();
+                    return extractionFn == null ? strValue : extractionFn.apply(strValue);
 
                   }
 
@@ -577,7 +579,7 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
                       if (dimIdx.length == 1) {
                         return dimDim.getValue(dimIdx[0]);
                       }
-                      String[] dimVals = new String[dimIdx.length];
+                      Comparable[] dimVals = new String[dimIdx.length];
                       for (int i = 0; i < dimIdx.length; i++) {
                         dimVals[i] = dimDim.getValue(dimIdx[i]);
                       }
@@ -592,6 +594,14 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
           }
         }
     );
+  }
+
+  private boolean isComparableNullOrEmpty(final Comparable value)
+  {
+    if (value instanceof String) {
+      return Strings.isNullOrEmpty((String) value);
+    }
+    return value == null;
   }
 
   private ValueMatcher makeFilterMatcher(final Filter filter, final EntryHolder holder)
@@ -638,18 +648,18 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
     }
 
     @Override
-    public ValueMatcher makeValueMatcher(String dimension, final String value)
+    public ValueMatcher makeValueMatcher(String dimension, final Comparable value)
     {
       IncrementalIndex.DimensionDesc dimensionDesc = index.getDimension(dimension);
       if (dimensionDesc == null) {
-        return new BooleanValueMatcher(Strings.isNullOrEmpty(value));
+        return new BooleanValueMatcher(isComparableNullOrEmpty(value));
       }
       final int dimIndex = dimensionDesc.getIndex();
       final IncrementalIndex.DimDim dimDim = dimensionDesc.getValues();
 
       final Integer id = dimDim.getId(value);
       if (id == null) {
-        if (Strings.isNullOrEmpty(value)) {
+        if (isComparableNullOrEmpty(value)) {
           return new ValueMatcher()
           {
             @Override
@@ -673,7 +683,7 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
         {
           int[][] dims = holder.getKey().getDims();
           if (dimIndex >= dims.length || dims[dimIndex] == null) {
-            return Strings.isNullOrEmpty(value);
+            return isComparableNullOrEmpty(value);
           }
 
           return Ints.indexOf(dims[dimIndex], id) >= 0;
@@ -682,7 +692,7 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
     }
 
     @Override
-    public ValueMatcher makeValueMatcher(String dimension, final Predicate<String> predicate)
+    public ValueMatcher makeValueMatcher(String dimension, final Predicate predicate)
     {
       IncrementalIndex.DimensionDesc dimensionDesc = index.getDimension(dimension);
       if (dimensionDesc == null) {
@@ -732,7 +742,12 @@ public class IncrementalIndexStorageAdapter implements StorageAdapter
           }
 
           for (int dimVal : dims[dimIndex]) {
-            List<String> stringCoords = Lists.newArrayList(SPLITTER.split(dimDim.getValue(dimVal)));
+            Comparable fullDimVal = dimDim.getValue(dimVal);
+            // TODO: decide what to do for non-String spatial dims, skip for now
+            if (!(fullDimVal instanceof String)) {
+              return false;
+            }
+            List<String> stringCoords = Lists.newArrayList(SPLITTER.split((String) fullDimVal));
             float[] coords = new float[stringCoords.size()];
             for (int j = 0; j < coords.length; j++) {
               coords[j] = Float.valueOf(stringCoords.get(j));