… benchmarks/src/main/java/io/druid/benchmark/IncrementalIndexAddRowsBenchmark.java
… processing/src/main/java/io/druid/segment/incremental/OnheapIncrementalIndex.java
… processing/src/main/java/io/druid/segment/incremental/SpatialDimensionRowTransformer.java
… server/src/test/java/io/druid/segment/realtime/plumber/RealtimePlumberSchoolTest.java