@jon-wei 2016-01-21
closed for now to rebase to master, will reopen