@jon-wei 2016-01-22
@drcrallen yeah, forgot about that yesterday =\

I was able to re-open it though by doing a "git reset --hard" to the SHA at the time the PR was closed