@fjy 2016-02-05
@jon-wei we need to rebase this