@jon-wei 2016-02-06
Updated this to rebase on top of changes introduced by:

https://github.com/druid-io/druid/pull/2085

The PR above changed TimeAndDims to use a dictionary encoded int[] array instead of String[]; this patch uses dictionary encoding for Long/Float values as well (unnecessary).

I plan to change TimeAndDims to use Comparable[] instead of int[] in a later PR, still WIP in my fork:
https://github.com/jon-wei/druid/tree/flex_dims_pt2

Strings will store a dictionary encoding in TimeAndDims as in PR #2085, numerics will store their values directly in TimeAndDims.