@nishantmonu51 2016-02-09
What is the  behavior of null for Long/Float dimensions ? 
Will it throw parseException or successfully ingest null values and treat them as null ? 

can we also add tests for null values in Long/Float dimension to make sure it behaves as expected.