@jon-wei 2016-02-09
@nishantmonu51 

RE: null handling, the nulls are being converted to 0.  It's not in this PR, but a follow on PR that I am finishing up:

https://github.com/jon-wei/druid/blob/5bffacec5772718bcb948b81e6ace4192d940396/processing/src/main/java/io/druid/segment/IndexMerger.java#L1432

That PR reuses the existing single-value long/float column formats for the numeric dims, representing null without changes to the storage format (need some mechanism to distinguish null from normal values) or using dictionary encoding (have a dictionary entry for null) seems problematic. Throwing an exception sounds okay to me too (maybe let user configure? choice of "throw exception" or "convert null to 0")

RE: docs and tests, this open PR doesn't actually enable the use of long/float dimensions, it's just preparing IncrementalIndex for later changes:
- It doesn't provide a way to specify dimension typing in the schema
- No query/filter/extraction fn, etc. support for long/floats yet

I address those points and have tests in the follow-on patch (not ready yet), I think it would make more sense to include docs in the later PR when the numeric dims are actually usable