@nishantmonu51 2016-02-11
+1, I think this is ready to be merged. docs can be added in the follow up PR.