@fjy 2016-02-19
@jon-wei do you mind resolving merge conflicts?